+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 - Fixed `echo` command appends `=` accidentially.
 - Added a very simple `curl` command.
+- Removed "exit code" feature.
+- Removed `true` command.
+- Removed `false` command.
 
 ## v0.1.0